[WIP MSG SPLIT] Add support for recent projects in the File menu.
This will be updated in the near future as I clean up a lot of the "project"
support.

Fixed a bug where Toplevel windows were being create a little down and to
the right of where they were supposed to.  This had to do with the wrong
window manager getting the geometry of the window.  It would not include the
title bar.

If the Console Window is killed by the OS window manager, we unset the console
button.

If you started a new project while in TEST mode, it would fail saying you
couldn't add widgets in TEST mode.  We now just set it to edit mode before
starting anew.

Added several functions to the right-click popup menu.
        Set Alias
        Cut
        Copy
        Paste
        Delete
        Copy Widgetname

Also added support for widget-specific options to the right-click menu.
Only a few of these are in place now.  I'll add more as we go.

Made the insertion of widgets more like traditional GUI builders.  When
selecting a widget from the toolbar, it depresses the button and waits for
the user to select an insertion point on the toplevel form.  When the user
clicks the form, the widget is created and placed at that point.

The insertion also takes into account the widget that the insertion is placed
upon.  So, if a point within a frame is clicked, the new widget is placed at
that point within the frame.

Compounds insert the same way as toolbar widgets.

Added an option to the preferences menu for turning off the new placement
of widgets. "Auto Place Widgets" will make widgets place the old way.

There are quite a number of other changes I can't think of right now.