[WIP MSG SPLIT] If you iconify/deiconify vTcl, it takes the toplevels with it.
Changed the default color for all vTcl windows to #d9d9d9.
Completely re-wrote the way widgets are defined.
Added widget command aliasing.
Added widget auto-aliasing
Added a pointer icon to the toolbar and continuous widget placement.
Added more options to the right-click menu.
Fixed a bug where clicking a Toplevel window in the widget tree
  would cause its manager to change, thereby breaking the project
  the next time it was sourced.
When a project is saved, we save the hidden toplevel windows too,
  we just don't show them again.  But, the next time we load the
  project, they're properly included in the project and on the widget
  tree.
Cleaned up the order in which widget selection occurs when a widget
  is deleted or cut.
Added a menu option to Restore a project from backup.