Removed type as a means of identifying widgets.  Widgets are all recognized by their class.
Re-wrote the core and vtcl widget files to reflect this change.

Changed the way LoadWidgets works.  Each subdirecty must be passed seperately.
This means we don't load widgets for libraries we don't support.

Modified all the system compounds to reflect the new type change.

If a window is withdrawn when we save, we add a withdraw command right
after the window's creation.  This let's us instatiate a window without
actually having to see it.

Added an update command when saving toplevels directly after the geometry
command.  This fixes a problem Tk has with resetting the geometry of a
withdrawn window to 1x1.

Scrollbars and Scales (arguably the hardest widgets to handle) now add their
toolbar icons by hand.  This takes a lot of pain out of handling seperate
widget types within a single widget class.