Merge of the 1.5 branch.
This one's a big one, and I don't even know if it's all intact.  I guess
we'll find out as we build.

Cleaned up the new tk_strictMotif code and created a MessageBox procedure
to do all the work.  This will give us a more uniform message box
throughout vTcl.