[SPLIT MULTI] xxx
Shift-click will now place the last widget placed.  This sort of bombs
the shift-click scrollbar feature, but since it's in the widget menu
of the scrollbar, we don't really lose functionality.

Removed the quit confirmation when exiting.

Add preference option for Widget Tree highlight color.

The command window remembers where it was and saves.